requirements.txt
# Ai-dev-team
# A multi-agent AI development team powered by crewAI

crewai>=0.80.0
crewai-tools>=0.20.0
langchain>=0.3.0
langchain-community>=0.3.0
openai>=1.0.0
anthropic>=0.25.0
chromadb>=0.5.0
tiktoken>=0.7.0
streamlit>=1.40.0
plotly>=5.24.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
rich>=13.0.0
pandas>=2.2.0